ping: Move the code a bit
This will allow adding subcommands for the ping handler.
